fix: reflect golangci-lint
