+++ x/wasm/keeper/exported_events.go
@@ -1,6 +1,15 @@
 package keeper
 
+import (
+	sdk "github.com/line/lbm-sdk/types"
+	wasmvmtypes "github.com/line/wasmvm/types"
+)
+
 var (
 	ContractSDKEventAttributes = contractSDKEventAttributes
 	EventTypeMinLength         = eventTypeMinLength
 )
+
+func NewCustomCallablePointEvents(evts wasmvmtypes.Events, contractAddr sdk.AccAddress, callstack []byte) (sdk.Events, error) {
+	return newCustomCallablePointEvents(evts, contractAddr, callstack)
+}
